+++ arm9/source/romBrowser/views/NdsGameDetailsBottomSheetView.cpp
@@ -16,10 +16,10 @@ NdsGameDetailsBottomSheetView::NdsGameDetailsBottomSheetView(
     , _cheatsChip(ChipView::CreateShared(md::sys::color::surfaceContainerLow, materialColorScheme, fontRepository))
     , _favoriteChip(ChipView::CreateShared(md::sys::color::surfaceContainerLow, materialColorScheme, fontRepository))
 {
-    _cheatsChip->SetText(u"Cheats");
+    _cheatsChip->SetText(u"金手指");
     _cheatsChip->SetSelected(false);
     AddChildTail(_cheatsChip.GetPointer());
-    _favoriteChip->SetText(u"Favorite");
+    _favoriteChip->SetText(u"收藏");
     _favoriteChip->SetSelected(true);
     AddChildTail(_favoriteChip.GetPointer());
 }