+++ arm9/source/romBrowser/views/deleteconfirm/DeleteConfirmBottomSheetView.cpp
@@ -27,17 +27,17 @@ DeleteConfirmBottomSheetView::DeleteConfirmBottomSheetView(SharedPtr<DeleteConfi
     , _hintLabel(Label2DView::CreateShared(LINE_WIDTH, 16, 40, fontRepository->GetFont(FontType::Medium7_5)))
     , _materialColorScheme(materialColorScheme)
 {
-    _titleLabel->SetText(u"Delete game?");
+    _titleLabel->SetText(u"删除游戏？");
     _fileNameLabel->SetEllipsisStyle(LabelView::EllipsisStyle::Marquee);
     _fileNameLabel->SetText(_viewModel->GetFileName());
     if (_viewModel->HasSave())
     {
         char text[280];
-        mini_snprintf(text, sizeof(text), "The save %s is also deleted", _viewModel->GetSaveFileName());
+        mini_snprintf(text, sizeof(text), "存档 %s 也会被删除", _viewModel->GetSaveFileName());
         _saveLabel->SetEllipsisStyle(LabelView::EllipsisStyle::Ellipsis);
         _saveLabel->SetText(text);
     }
-    _hintLabel->SetText("X: delete    A/B: cancel");
+    _hintLabel->SetText("X: 删除    A/B: 取消");
     AddChildTail(_titleLabel.GetPointer());
     AddChildTail(_fileNameLabel.GetPointer());
     if (_viewModel->HasSave())