+++ arm9/source/themes/DefaultFontRepository.cpp
@@ -1,8 +1,5 @@
 #include "common.h"
-#include "NotoSansJP-Regular-10_nft2.h"
-#include "NotoSansJP-Medium-7_5_nft2.h"
-#include "NotoSansJP-Medium-10_nft2.h"
-#include "NotoSansJP-Medium-11_nft2.h"
+#include "BitmapSong-9pt_nft2.h"
 #include "DefaultFontRepository.h"
 
 const nft2_header_t* DefaultFontRepository::GetFont(FontType fontType) const
@@ -11,19 +8,19 @@ const nft2_header_t* DefaultFontRepository::GetFont(FontType fontType) const
     {
         case FontType::Regular10:
         {
-            return (const nft2_header_t*)NotoSansJP_Regular_10_nft2;
+            return (const nft2_header_t*)BitmapSong_9pt_nft2;
         }
         case FontType::Medium7_5:
         {
-            return (const nft2_header_t*)NotoSansJP_Medium_7_5_nft2;
+            return (const nft2_header_t*)BitmapSong_9pt_nft2;
         }
         case FontType::Medium10:
         {
-            return (const nft2_header_t*)NotoSansJP_Medium_10_nft2;
+            return (const nft2_header_t*)BitmapSong_9pt_nft2;
         }
         case FontType::Medium11:
         {
-            return (const nft2_header_t*)NotoSansJP_Medium_11_nft2;
+            return (const nft2_header_t*)BitmapSong_9pt_nft2;
         }
         default:
         {